target/i386: sev: Return 0 if sev_send_get_packet_len() fails
The send_packet_hdr_len of struct SEVState is of type size_t
which is an unsigned class type. If the send_packet_hdr_len
is assigned as -1, then it will be a huge number and the QEMU
process will crash when allocating packet buffer with the
huge size.

For example, the following code could cause crash described
above.

    ```
    static int
    sev_send_update_data(SEVState *s, QEMUFile *f, uint8_t *ptr, uint32_t size,
                     uint64_t *bytes_sent)
    {

        ......

        if (!s->send_packet_hdr) {
            s->send_packet_hdr_len = sev_send_get_packet_len(&fw_error);
            if (s->send_packet_hdr_len < 1) {
                error_report("%s: SEND_UPDATE fw_error=%d '%s'",
                             __func__, fw_error, fw_error_to_str(fw_error));
                return 1;
            }

            s->send_packet_hdr = g_new(gchar, s->send_packet_hdr_len);
        }

        ......

    }
    ```

Signed-off-by: hanliyang <hanliyang@hygon.cn>